golang: Cast correctly for portability to 32 bit armv7.
../../src/libguestfs.org/nbdkit/nbdkit.go:532:22: cannot use struct_size (type _Ctype_ulong) as type _Ctype_ulonglong in assignment
../../src/libguestfs.org/nbdkit/nbdkit.go:566:41: cannot use struct_size (type _Ctype_ulong) as type _Ctype_uint in argument to _Cfunc__CMalloc

(cherry picked from commit 44571561854f9c2cde2bb0e55e430c19eb866cb9)

+++ plugins/golang/src/libguestfs.org/nbdkit/nbdkit.go
@@ -529,7 +529,7 @@ func PluginInitialize(name string, impl PluginInterface) unsafe.Pointer {
 
 	// Set up the hidden plugin fields as for C.
 	struct_size := C.ulong(unsafe.Sizeof(plugin))
-	plugin._struct_size = struct_size
+	plugin._struct_size = C.uint64_t(struct_size)
 	plugin._api_version = C.NBDKIT_API_VERSION
 	plugin._thread_model = C.NBDKIT_THREAD_MODEL_PARALLEL
 
@@ -563,7 +563,7 @@ func PluginInitialize(name string, impl PluginInterface) unsafe.Pointer {
 	// Return a newly malloced copy of the struct.  This must be
 	// globally available to the C code in the server, so it is
 	// never freed.
-	p := (*C.struct_nbdkit_plugin)(C.malloc(struct_size))
+	p := (*C.struct_nbdkit_plugin)(C.malloc(C.size_t(struct_size)))
 	*p = plugin
 	return unsafe.Pointer(p)
 }
